mejora de la base de datos y creacion de la capa logica de negocios de ventas y arreglos de cln producto y clientes

+++ Hamburgueseria/CadHamburgueseria/DetallePedido.cs
@@ -17,7 +17,7 @@ namespace CadHamburgueseria
         public int id { get; set; }
         public int idPedido { get; set; }
         public int idProducto { get; set; }
-        public decimal cantidad { get; set; }
+        public int cantidad { get; set; }
         public decimal precioUnitario { get; set; }
         public Nullable<decimal> total { get; set; }
         public Nullable<short> estado { get; set; }
